+++ test/tests/SubSpeciesTest.C
@@ -50,7 +50,7 @@ TEST(SubSpecies, PhotonTest)
   EXPECT_EQ(s.getName(), "hnu");
   EXPECT_EQ(s.getBase(), "hnu");
   EXPECT_EQ(s.getModifier(), "");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["hnu"]);
   EXPECT_EQ(s.getChargeNumber(), 0);
   EXPECT_FLOAT_EQ(s.getCharge(), 0);
@@ -64,7 +64,7 @@ TEST(SubSpecies, LowerCasesElectronTest)
   EXPECT_EQ(s.getName(), "e");
   EXPECT_EQ(s.getBase(), "e");
   EXPECT_EQ(s.getModifier(), "");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -1);
   EXPECT_FLOAT_EQ(s.getCharge(), -e);
@@ -78,7 +78,7 @@ TEST(SubSpecies, UpperCasesElectronTest)
   EXPECT_EQ(s.getName(), "E");
   EXPECT_EQ(s.getBase(), "E");
   EXPECT_EQ(s.getModifier(), "");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -1);
   EXPECT_FLOAT_EQ(s.getCharge(), -e);
@@ -92,7 +92,7 @@ TEST(SubSpecies, GroundStateNoCharge)
   EXPECT_EQ(s.getName(), "Ar");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["Ar"]);
   EXPECT_EQ(s.getChargeNumber(), 0);
   EXPECT_FLOAT_EQ(s.getCharge(), 0);
@@ -106,7 +106,7 @@ TEST(SubSpecies, GroundStateSinglePositiveIon)
   EXPECT_EQ(s.getName(), "Ar+");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "+");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["Ar"] - base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), 1);
   EXPECT_FLOAT_EQ(s.getCharge(), e);
@@ -120,7 +120,7 @@ TEST(SubSpecies, GroundStateMultiplePositiveIon)
   EXPECT_EQ(s.getName(), "Ar+4");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "+4");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["Ar"] - 4 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), 4);
   EXPECT_FLOAT_EQ(s.getCharge(), 4 * e);
@@ -134,7 +134,7 @@ TEST(SubSpecies, GroundStateSingleNegativeIon)
   EXPECT_EQ(s.getName(), "Ar-");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "-");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["Ar"] + base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -1);
   EXPECT_FLOAT_EQ(s.getCharge(), -e);
@@ -148,7 +148,7 @@ TEST(SubSpecies, GroundStateMultipleNegativeIon)
   EXPECT_EQ(s.getName(), "Ar-100");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "-100");
-  EXPECT_EQ(s.getSubscript(), 1);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 1);
   EXPECT_FLOAT_EQ(s.getMass(), base_masses["Ar"] + 100 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -100);
   EXPECT_FLOAT_EQ(s.getCharge(), -100 * e);
@@ -162,7 +162,7 @@ TEST(SubSpecies, GroundStateMolecular)
   EXPECT_EQ(s.getName(), "Ar2");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "2");
-  EXPECT_EQ(s.getSubscript(), 2);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 2);
   EXPECT_FLOAT_EQ(s.getMass(), 2 * base_masses["Ar"]);
   EXPECT_EQ(s.getChargeNumber(), 0);
   EXPECT_FLOAT_EQ(s.getCharge(), 0);
@@ -176,7 +176,7 @@ TEST(SubSpecies, GroundStateLargeMolecule)
   EXPECT_EQ(s.getName(), "Ar188");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "188");
-  EXPECT_EQ(s.getSubscript(), 188);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 188);
   EXPECT_FLOAT_EQ(s.getMass(), 188 * base_masses["Ar"]);
   EXPECT_EQ(s.getChargeNumber(), 0);
   EXPECT_FLOAT_EQ(s.getCharge(), 0);
@@ -190,7 +190,7 @@ TEST(SubSpecies, MolecularNegativeIon)
   EXPECT_EQ(s.getName(), "Ar2-100");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "2-100");
-  EXPECT_EQ(s.getSubscript(), 2);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 2);
   EXPECT_FLOAT_EQ(s.getMass(), 2 * base_masses["Ar"] + 100 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -100);
   EXPECT_FLOAT_EQ(s.getCharge(), -100 * e);
@@ -204,7 +204,7 @@ TEST(SubSpecies, MolecularPositiveIon)
   EXPECT_EQ(s.getName(), "Ar120+39");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "120+39");
-  EXPECT_EQ(s.getSubscript(), 120);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 120);
   EXPECT_FLOAT_EQ(s.getMass(), 120 * base_masses["Ar"] - 39 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), 39);
   EXPECT_FLOAT_EQ(s.getCharge(), 39 * e);
@@ -218,7 +218,7 @@ TEST(SubSpecies, MolecularPositiveIonWithModifier)
   EXPECT_EQ(s.getName(), "Ar120+39(test)");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "120+39(test)");
-  EXPECT_EQ(s.getSubscript(), 120);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 120);
   EXPECT_FLOAT_EQ(s.getMass(), 120 * base_masses["Ar"] - 39 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), 39);
   EXPECT_FLOAT_EQ(s.getCharge(), 39 * e);
@@ -232,7 +232,7 @@ TEST(SubSpecies, MolecularNegativeIonWithModifier)
   EXPECT_EQ(s.getName(), "Ar2-100(test)");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "2-100(test)");
-  EXPECT_EQ(s.getSubscript(), 2);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 2);
   EXPECT_FLOAT_EQ(s.getMass(), 2 * base_masses["Ar"] + 100 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -100);
   EXPECT_FLOAT_EQ(s.getCharge(), -100 * e);
@@ -247,7 +247,7 @@ TEST(SubSpecies, CustomSpeciesBase)
   EXPECT_EQ(s.getName(), "A2-100(test)");
   EXPECT_EQ(s.getBase(), "A");
   EXPECT_EQ(s.getModifier(), "2-100(test)");
-  EXPECT_EQ(s.getSubscript(), 2);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 2);
   EXPECT_FLOAT_EQ(s.getMass(), 2 * base_masses["A"] + 100 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -100);
   EXPECT_FLOAT_EQ(s.getCharge(), -100 * e);
@@ -263,7 +263,7 @@ TEST(SubSpecies, MassOverride)
   EXPECT_EQ(s.getName(), "Ar2-100(test)");
   EXPECT_EQ(s.getBase(), "Ar");
   EXPECT_EQ(s.getModifier(), "2-100(test)");
-  EXPECT_EQ(s.getSubscript(), 2);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 2);
   EXPECT_FLOAT_EQ(s.getMass(), 2 * 5 + 100 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -100);
   EXPECT_FLOAT_EQ(s.getCharge(), -100 * e);
@@ -278,7 +278,7 @@ TEST(SubSpecies, LongCustomSpecies)
   EXPECT_EQ(s.getName(), "Polypeptide2-100(test)");
   EXPECT_EQ(s.getBase(), "Polypeptide");
   EXPECT_EQ(s.getModifier(), "2-100(test)");
-  EXPECT_EQ(s.getSubscript(), 2);
+  EXPECT_EQ(s.getSubscript(), (unsigned int) 2);
   EXPECT_FLOAT_EQ(s.getMass(), 2 * 1000 + 100 * base_masses["e"]);
   EXPECT_EQ(s.getChargeNumber(), -100);
   EXPECT_FLOAT_EQ(s.getCharge(), -100 * e);
